Tuple Unpacking Functions
Using functions, we can unpack tuples and process them through for loops too. This lecture, I successfully managed to unpack tuples via function/loops and using if statements to meet conditions which output the data we are searching for. Then I was able to dissect the data to retrieve specific information and unpacking the tuple to utilise it later for whatever reason.