Interaction Between functions
Funception stuff. Made a mini guess the ball location game via multiple functions being called to build the and imitate games slightly complex but overall simple logic in python code. Learned the new random library method called shuffle. Shuffle plays a key role for the game as it's the essence of the game.